+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+ - Tooltip and Description component settings to be Ace Textarea so that it would be convenient for HTML
+
 ## 4.8.0-rc.14
 ### Changed
  - Added improvements to accessibility.