+++ Changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+ - Ace Textarea issues
+   - Placeholder not disappearing when typing
+
 ## 4.8.0-rc.14
 ### Changed
  - Added improvements to accessibility.