+++ Changelog.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
  - Tooltip and Description component settings to be Ace Textarea so that it would be convenient for HTML
 
+### Fixed
+ - ```userPermissions``` method not handling submissions with multiple groups properly 
+
 ## 4.8.0-rc.14
 ### Changed
  - Added improvements to accessibility.