+++ Changelog.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
    - Editor settings from component JSON being ignored
    - `minLines`, `maxLines`, `tabSize` Aces settings being hardcoded without ability to override
 
+### Changed
+ - Tooltip and Description component settings to be Ace Textarea so that it would be convenient for HTML
+
+### Fixed
+ - ```userPermissions``` method not handling submissions with multiple groups properly 
+
 ## 4.8.0-rc.14
 ### Changed
  - Added improvements to accessibility.